# meterapi/Data/Mappers/NewMeterViewModel.cs
using System.ComponentModel.DataAnnotations;

namespace meterapi.Data.Mappers
{
    public class NewMeterViewModel
    {
       
        public string MeterDeviceId { get; set; }
        public string RpId { get; set; }
    }
}