# meterapi/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "meterapi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Email/verify/{token}": {
      "post": {
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Email/verify": {
      "post": {
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Email/forgotpassword": {
      "post": {
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Email/resetpassword/{token}": {
      "post": {
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "newPassword",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Meter/{id}": {
      "get": {
        "tags": [
          "Meter"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Meter"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/NewMeterViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Meter"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/MeterData": {
      "get": {
        "tags": [
          "MeterData"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MeterData"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MeterData"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MeterDataDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/MeterData"
            }
          }
        }
      }
    },
    "/api/MeterData/{id}": {
      "get": {
        "tags": [
          "MeterData"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/MeterData"
            }
          }
        }
      },
      "put": {
        "tags": [
          "MeterData"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MeterDataDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "MeterData"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/MeterData"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Meter": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "meterDeviceId": {
          "type": "string"
        },
        "rpId": {
          "type": "string"
        },
        "meterDatas": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MeterData"
          }
        },
        "userMeter": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserMeter"
          }
        }
      }
    },
    "MeterData": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "meterId": {
          "format": "int32",
          "type": "integer"
        },
        "totalConsumptionDay": {
          "format": "float",
          "type": "number"
        },
        "totalConsumptionNight": {
          "format": "float",
          "type": "number"
        },
        "allPhaseConsumption": {
          "format": "float",
          "type": "number"
        },
        "gasConsumption": {
          "format": "float",
          "type": "number"
        },
        "meter": {
          "$ref": "#/definitions/Meter"
        }
      }
    },
    "MeterDataDTO": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "meterId": {
          "format": "int32",
          "type": "integer"
        },
        "totalConsumptionDay": {
          "format": "float",
          "type": "number"
        },
        "totalConsumptionNight": {
          "format": "float",
          "type": "number"
        },
        "allPhaseConsumption": {
          "format": "float",
          "type": "number"
        },
        "gasConsumption": {
          "format": "float",
          "type": "number"
        }
      }
    },
    "NewMeterViewModel": {
      "type": "object",
      "properties": {
        "meterDeviceId": {
          "type": "string"
        },
        "rpId": {
          "type": "string"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "passwordHash": {
          "format": "byte",
          "type": "string"
        },
        "passwordSalt": {
          "format": "byte",
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "tokenCreated": {
          "format": "date-time",
          "type": "string"
        },
        "tokenExpires": {
          "format": "date-time",
          "type": "string"
        },
        "userMeters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserMeter"
          }
        },
        "verificationToken": {
          "type": "string",
          "readOnly": true
        },
        "isEmailVerified": {
          "type": "boolean"
        },
        "resetPasswordToken": {
          "type": "string",
          "readOnly": true
        },
        "resetPasswordExpiration": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        }
      }
    },
...
    "UserMeter": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "meterId": {
          "format": "int32",
          "type": "integer"
        },
        "rpId": {
          "type": "string"
        },
        "meterDeviceId": {
          "type": "string"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "address": {
          "type": "string"
        },
        "meter": {
          "$ref": "#/definitions/Meter"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
  }
}